refactor: add config/backoff.go
